Corrected MatrixAverageCalculator

+++ src/main/java/study/base/arrays/matrix/MatrixAverageCalculator.java
@@ -11,14 +11,14 @@ package study.base.arrays.matrix;
 public class MatrixAverageCalculator {
 
   public double calculate(int[][] matrix) {
-    int sumOfNumber = 0;
-    int matrixSumLenght = 0;
+    double sumOfNumber = 0;
+    double matrixSumLenght = 0;
     for (int [] matrixLine : matrix){
       for (int matrixNumb : matrixLine){
         sumOfNumber += matrixNumb;
         matrixSumLenght++;
       }
     }
-    return (double) sumOfNumber / matrixSumLenght;
+    return sumOfNumber / matrixSumLenght;
   }
 }
